fixing the tests for the endpoint rename
